Address last comment
