+++ torch_points3d/core/losses/__init__.py
@@ -1,6 +1,7 @@
 import sys
 
 from .losses import *
+from .huber_loss import *
 from pytorch_metric_learning.miners import *
 from pytorch_metric_learning.losses import *
