+++ torch_points3d/core/losses/huber_loss.py
@@ -1,0 +1,30 @@
+import torch
+
+
+def huber_loss(error, delta=1.0):
+    """
+    Args:
+        error: Torch tensor (d1,d2,...,dk)
+    Returns:
+        loss: Torch tensor (d1,d2,...,dk)
+
+    x = error = pred - gt or dist(pred,gt)
+    0.5 * |x|^2                 if |x|<=d
+    0.5 * d^2 + d * (|x|-d)     if |x|>d
+    Ref: https://github.com/charlesq34/frustum-pointnets/blob/master/models/model_util.py
+    """
+    abs_error = torch.abs(error)
+    # quadratic = torch.min(abs_error, torch.FloatTensor([delta]))
+    quadratic = torch.clamp(abs_error, max=delta)
+    linear = abs_error - quadratic
+    loss = 0.5 * quadratic ** 2 + delta * linear
+    return loss
+
+
+class HuberLoss(torch.nn.Module):
+    def __init__(self, delta=0.1):
+        super().__init__()
+        self._delta = delta
+
+    def forward(self, error):
+        return huber_loss(error, self._delta)
